+++ frontend/src/components/Hazard/SiteSelection/SiteSelectionForm.js
@@ -3,6 +3,7 @@ import React, { Fragment, useContext, useEffect, useState } from "react";
 import { v4 as uuidv4 } from "uuid";
 import { GlobalContext } from "context";
 import { useAuth0 } from "components/common/ReactAuth0SPA";
+import GuideTooltip from "components/common/GuideTooltip";
 import * as CONSTANTS from "constants/Constants";
 import { disableScrollOnNumInput, handleErrors, sortIMs } from "utils/Utils";
 import TextField from "@material-ui/core/TextField";
@@ -292,7 +293,7 @@ const SiteSelectionForm = () => {
     <Fragment>
       {CONSTANTS.ENV === "DEV" ? (
         <div>
-          <div className="form-row form-section-title">
+          <div className="form-group form-section-title">
             <span>Ensemble</span>
           </div>
           <div className="custom-form-group">
@@ -301,7 +302,12 @@ const SiteSelectionForm = () => {
         </div>
       ) : null}
 
-      <div className="form-row form-section-title">Location</div>
+      <div className="form-group form-section-title">
+        Location
+        <GuideTooltip
+          explanation={CONSTANTS.TOOLTIP_MESSAGES["SITE_SELECTION_LOCATION"]}
+        />
+      </div>
 
       <form autoComplete="off" onSubmit={(e) => e.preventDefault()}>
         <div className="form-group">
@@ -387,7 +393,14 @@ const SiteSelectionForm = () => {
         </div>
       </form>
 
-      <div className="form-row form-section-title">Site Conditions</div>
+      <div className="form-group form-section-title">
+        Site Conditions
+        <GuideTooltip
+          explanation={
+            CONSTANTS.TOOLTIP_MESSAGES["SITE_SELECTION_SITE_CONDITION"]
+          }
+        />
+      </div>
 
       <SiteConditions />
     </Fragment>