+++ frontend/src/components/Project/SeismicHazard/HazardCurveSection.js
@@ -3,7 +3,9 @@ import React, { useState, useContext, useEffect, Fragment } from "react";
 import { v4 as uuidv4 } from "uuid";
 import { GlobalContext } from "context";
 
+import * as CONSTANTS from "constants/Constants";
 import ProjectSelect from "components/common/ProjectSelect";
+import GuideTooltip from "components/common/GuideTooltip";
 
 const HazardCurveSection = () => {
   const {
@@ -36,7 +38,10 @@ const HazardCurveSection = () => {
   return (
     <Fragment>
       <div className="form-group form-section-title">
-        <span>Hazard Curve</span>
+        Hazard Curve
+        <GuideTooltip
+          explanation={CONSTANTS.TOOLTIP_MESSAGES["PROJECT_HAZARD"]}
+        />
       </div>
       <div className="form-group">
         <label