+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## February 3, 2021
+
+### Tooltips - ([PR #40](https://github.com/ucgmsim/seistech_psha_frontend/pull/40))
+
+- Tooltips with a brief explanation.
+- Disable the MapBox in Hazard Analysis temporarily due to the crash.
+
 ## January 27, 2021
 
 ### Filter UHS section based on given IM list - ([PR #38](https://github.com/ucgmsim/seistech_psha_frontend/pull/38))